created routers and optimized usage
