+++ todo_app/database.py
@@ -3,7 +3,9 @@ Going to be used for link FastAPI application with our sqlite database
 """
 from sqlalchemy import create_engine
 from sqlalchemy.orm import sessionmaker
-from sqlalchemy.ext.declarative import declarative_base
+
+# from sqlalchemy.ext.declarative import declarative_base
+from sqlalchemy.orm import declarative_base
 
 SQLALCHEMY_DATABASE_URL = "sqlite:///./todo.db"
