+++ todo_app/routers/auth.py
@@ -1,0 +1,8 @@
+from fastapi import APIRouter
+
+router = APIRouter()
+
+
+@router.get("/auth")
+def get_user():
+    return {"user": "authenticated"}
